_BACKUP_MD/TES_Manual_Digital/TES_Manual_Digital/requirements_word.txt
# Requisitos para conversión de Markdown a Word
# Instalar con: pip install -r requirements_word.txt

# Opción 1: Pandoc (recomendado - instalación del sistema)
# sudo apt install pandoc  # En Ubuntu/Debian
# brew install pandoc      # En macOS
# choco install pandoc     # En Windows (con Chocolatey)

# Opción 2: Bibliotecas Python (alternativa si pandoc no está disponible)
python-docx>=0.8.11
markdown>=3.4.0
markdown-extensions>=0.6.0

# Opcionales (mejor calidad de conversión)
beautifulsoup4>=4.11.0  # Para parsing HTML
lxml>=4.9.0             # Parser XML para BeautifulSoup
